+++ src/RecoveryServices/RecoveryServices.Backup.Test/ScenarioTests/IaasVm/ItemTests.ps1
@@ -12,6 +12,58 @@
 # limitations under the License.
 # ----------------------------------------------------------------------------------
 
+function Test-AzureMonitorAlerts
+{
+	$location = "centraluseuap"
+	$resourceGroupName = "hiagarg"
+	$vaultName1 = "alerts-pstest-vault1"
+	$vaultName2 = "alerts-pstest-vault2"
+
+	try
+	{	
+		# create a vault without Alert settings
+		$tag= @{"MABUsed"="Yes";"Owner"="hiaga";"Purpose"="Testing";"DeleteBy"="06-2099"}
+		$vault1 = New-AzRecoveryServicesVault -Name $vaultName1 -ResourceGroupName $resourceGroupName -Location "centraluseuap" -Tag $tag
+		
+		Assert-True { $vault1.Properties.AlertSettings -eq $null }
+
+		# create a vault with Alert settings 
+		$vault2 = New-AzRecoveryServicesVault -Name $vaultName2 -ResourceGroupName $resourceGroupName -Location "centraluseuap" `
+			-Tag $tag -DisableAzureMonitorAlertsForJobFailure $false `
+			-DisableClassicAlerts $true			
+		
+		Assert-True { $vault2.Properties.AlertSettings -ne $null }
+		Assert-True { $vault2.Properties.AlertSettings.AzureMonitorAlertsForAllJobFailure -eq "Enabled" }
+		Assert-True { $vault2.Properties.AlertSettings.ClassicAlertsForCriticalOperations -eq "Disabled" }
+
+		$vault = Update-AzRecoveryServicesVault -ResourceGroupName "hiagarg"  -Name "hiagaVault" -DisableClassicAlerts $false
+
+		# update alert settings 
+		$vault1 = Update-AzRecoveryServicesVault -Name $vaultName1 -ResourceGroupName $resourceGroupName `
+			-DisableAzureMonitorAlertsForJobFailure $false `
+			-DisableClassicAlerts $true
+
+		Assert-True { $vault1.Properties.AlertSettings -ne $null }
+		Assert-True { $vault1.Properties.AlertSettings.AzureMonitorAlertsForAllJobFailure -eq "Enabled" }
+		Assert-True { $vault1.Properties.AlertSettings.ClassicAlertsForCriticalOperations -eq "Disabled" }
+		
+		$vault2 = Update-AzRecoveryServicesVault -Name $vaultName2 -ResourceGroupName $resourceGroupName `
+			-DisableAzureMonitorAlertsForJobFailure $true `
+			-DisableClassicAlerts $false
+
+		Assert-True { $vault2.Properties.AlertSettings -ne $null }
+		Assert-True { $vault2.Properties.AlertSettings.AzureMonitorAlertsForAllJobFailure -eq "Disabled" }
+		Assert-True { $vault2.Properties.AlertSettings.ClassicAlertsForCriticalOperations -eq "Enabled" }
+
+	}
+	finally
+	{
+		# Cleanup
+		Remove-AzRecoveryServicesVault -Vault $vault1
+		Remove-AzRecoveryServicesVault -Vault $vault2
+	}
+}
+
 function Test-AzureVMMUA
 {
 	$location = "centraluseuap"
@@ -196,8 +248,8 @@ function Test-AzureVMRestoreWithMSI
 function Test-AzureVMCrossRegionRestore
 {
 	$location = "centraluseuap"
-	$resourceGroupName = Create-ResourceGroup $location 24
-
+	$resourceGroupName = Create-ResourceGroup $location 24			
+	
 	try
 	{
 		# Setup
@@ -313,21 +365,29 @@ function Test-AzureBackupDataMove
 
 function Test-AzureVMGetItems
 {
-	$location = "southeastasia"
-	$resourceGroupName = Create-ResourceGroup $location
+	$location = "centraluseuap"
+	$resourceGroupName = "hiagarg"
+	$vaultName = "hiaga-adhoc-vault"
+	$vmName1 = "VM;iaasvmcontainerv2;hiagarg;hiaga-adhoc-vm"
+	$vmName2 = "VM;iaasvmcontainerv2;hiagarg;hiaganewvm3"	
+	$vmFriendlyName1 = "hiaga-adhoc-vm"
+	$vmFriendlyName2 = "hiaganewvm3"
+	$protectionState = "IRPending"
+	#$location = "southeastasia"
+	#$resourceGroupName = Create-ResourceGroup $location
 	
 	try
 	{
 		# Setup
-		$vm = Create-VM $resourceGroupName $location 1
-		$vm2 = Create-VM $resourceGroupName $location 12
-		$vault = Create-RecoveryServicesVault $resourceGroupName $location
+		$vm = Get-AzVM -ResourceGroupName $resourceGroupName -Name $vmFriendlyName1
+		$vm2 = Get-AzVM -ResourceGroupName $resourceGroupName -Name $vmFriendlyName2
+		$vault = Get-AzRecoveryServicesVault -ResourceGroupName $resourceGroupName -Name $vaultName 
 
 		# disable soft delete for successful cleanup
 		Set-AzRecoveryServicesVaultProperty -VaultId $vault.ID -SoftDeleteFeatureState "Disable"
 
-		Enable-Protection $vault $vm
-		Enable-Protection $vault $vm2
+		# Enable-Protection $vault $vm
+		# Enable-Protection $vault $vm2
 		$policy = Get-AzRecoveryServicesBackupProtectionPolicy `
 			-VaultId $vault.ID `
 			-Name "DefaultPolicy"
@@ -369,7 +429,7 @@ function Test-AzureVMGetItems
 			-VaultId $vault.ID `
 			-Container $container `
 			-WorkloadType AzureVM `
-			-ProtectionState IRPending;
+			-ProtectionState Protected;
 		Assert-True { $items.VirtualMachineId -contains $vm.Id }
 
 		# VARIATION-5: Get items for container with friendly name and ProtectionStatus filters
@@ -387,7 +447,7 @@ function Test-AzureVMGetItems
 			-Container $container `
 			-WorkloadType AzureVM `
 			-Name $vm.Name `
-			-ProtectionState IRPending;
+			-ProtectionState Protected;
 		Assert-True { $items.VirtualMachineId -contains $vm.Id }
 
 		# VARIATION-7: Get items for container with Status and ProtectionStatus filters
@@ -395,7 +455,7 @@ function Test-AzureVMGetItems
 			-VaultId $vault.ID `
 			-Container $container `
 			-WorkloadType AzureVM `
-			-ProtectionState IRPending `
+			-ProtectionState Protected `
 			-ProtectionStatus Healthy;
 		Assert-True { $items.VirtualMachineId -contains $vm.Id }
 
@@ -405,7 +465,7 @@ function Test-AzureVMGetItems
 			-Container $container `
 			-WorkloadType AzureVM `
 			-Name $vm.Name `
-			-ProtectionState IRPending `
+			-ProtectionState Protected `
 			-ProtectionStatus Healthy;
 		Assert-True { $items.VirtualMachineId -contains $vm.Id }
 
@@ -417,30 +477,48 @@ function Test-AzureVMGetItems
 	}
 	finally
 	{
-		# Cleanup
-		Cleanup-ResourceGroup $resourceGroupName		
+		# Cleanup				
 	}
 }
 
 function Test-AzureVMProtection
 {
-	$location = "southeastasia"
-	$resourceGroupName = Create-ResourceGroup $location
+	# $location = "southeastasia"
+	# $resourceGroupName = Create-ResourceGroup $location
+
+	$location = "centraluseuap"
+	$resourceGroupName = "iaasvm-pstest-rg"
+	$vaultName = "iaasvm-pstest-vault"
+	$vmName = "iaasvm-pstest-vm"
 
 	try
 	{
 		# Setup
-		$vm = Create-VM $resourceGroupName $location
-		$vault = Create-RecoveryServicesVault $resourceGroupName $location
+		$vm = Get-AzVM -ResourceGroupName $resourceGroupName -Name $vmName
+		$vault = Get-AzRecoveryServicesVault -ResourceGroupName $resourceGroupName -Name $vaultName
+		
 		Set-AzRecoveryServicesVaultProperty -VaultId $vault.ID -SoftDeleteFeatureState "Disable"
-		# Sleep to give the service time to add the default policy to the vault
-        Start-TestSleep 5000
 
-		# Get default policy
+		$item = Get-AzRecoveryServicesBackupItem -BackupManagementType AzureVM -WorkloadType AzureVM -VaultId $vault.ID
+
+		Disable-AzRecoveryServicesBackupProtection `
+			-VaultId $vault.ID `
+			-Item $item `
+			-RemoveRecoveryPoints `
+			-Force;
+
 		$policy = Get-AzRecoveryServicesBackupProtectionPolicy `
 			-VaultId $vault.ID `
 			-Name "DefaultPolicy";
-	
+
+		Assert-True {$policy.ProtectedItemsCount -eq 0};
+
+		# $vm = Create-VM $resourceGroupName $location
+		# $vault = Create-RecoveryServicesVault $resourceGroupName $location
+		
+		# Sleep to give the service time to add the default policy to the vault
+        Start-TestSleep 5000
+
 		# Enable protection
 		Enable-AzRecoveryServicesBackupProtection `
 			-VaultId $vault.ID `
@@ -454,49 +532,36 @@ function Test-AzureVMProtection
 
 		Assert-True {$policy.ProtectedItemsCount -eq 1};
 
-		$container = Get-AzRecoveryServicesBackupContainer `
-			-VaultId $vault.ID `
-			-ContainerType AzureVM `
-			-Status Registered;
-
-		$item = Get-AzRecoveryServicesBackupItem `
-			-VaultId $vault.ID `
-			-Container $container `
-			-WorkloadType AzureVM
-
-		# Disable protection
-		Disable-AzRecoveryServicesBackupProtection `
-			-VaultId $vault.ID `
-			-Item $item `
-			-RemoveRecoveryPoints `
-			-Force;
+		$item = Get-AzRecoveryServicesBackupItem -BackupManagementType AzureVM -WorkloadType AzureVM -VaultId $vault.ID
 		
-		$policy = Get-AzRecoveryServicesBackupProtectionPolicy `
-			-VaultId $vault.ID `
-			-Name "DefaultPolicy";
-
-		Assert-True {$policy.ProtectedItemsCount -eq 0};
-
+		Assert-True { $item.SourceResourceId -match $vm.Name };
 	}
 	finally
 	{
 		# Cleanup
-		Cleanup-ResourceGroup $resourceGroupName
+		Set-AzRecoveryServicesVaultProperty -VaultId $vault.ID -SoftDeleteFeatureState "Enable"
 	}
 }
 
 function Test-AzureVMGetRPs
-{
-	$location = "southeastasia"
-	$resourceGroupName = Create-ResourceGroup $location
+{	
+	$location = "centraluseuap"
+	$resourceGroupName = "iaasvm-pstest-rg"
+	$vaultName = "iaasvm-pstest-vault"
+	$vmName = "iaasvm-pstest-vm"
+
+	# $location = "southeastasia"
+	# $resourceGroupName = Create-ResourceGroup $location
 
 	try
 	{
   		# Setup
-		$vm = Create-VM $resourceGroupName $location
-		$vault = Create-RecoveryServicesVault $resourceGroupName $location
+		$vm = Get-AzVM -ResourceGroupName $resourceGroupName -Name $vmName #Create-VM $resourceGroupName $location
+		$vault = Get-AzRecoveryServicesVault -ResourceGroupName $resourceGroupName -Name $vaultName # Create-RecoveryServicesVault $resourceGroupName $location
+		
 		Set-AzRecoveryServicesVaultProperty -VaultId $vault.ID -SoftDeleteFeatureState "Disable"
-		$item = Enable-Protection $vault $vm
+
+		$item = Get-AzRecoveryServicesBackupItem -BackupManagementType AzureVM -WorkloadType AzureVM -VaultId $vault.ID # Enable-Protection $vault $vm
 		$backupJob = Backup-Item $vault $item
 
 		# Test 1: Get latest recovery point; should be only one
@@ -547,24 +612,41 @@ function Test-AzureVMGetRPs
 	finally
 	{
 		# Cleanup
-		Cleanup-ResourceGroup $resourceGroupName
+		# Cleanup-ResourceGroup $resourceGroupName
+		
+		# disable protection with RemoveRecoveryPoints
+		# Disable-AzRecoveryServicesBackupProtection -Item $item -RemoveRecoveryPoints -VaultId $vault.ID -Force
+
+		# enable soft delete 
+		Set-AzRecoveryServicesVaultProperty -SoftDeleteFeatureState Enable -VaultId $vault.ID
 	}
 }
 
 function Test-AzureVMFullRestore
 {
-	$location = "southeastasia"
-	$resourceGroupName = Create-ResourceGroup $location
-	$targetResourceGroupName = Create-ResourceGroup $location 1
+	# $location = "southeastasia"
+	# $resourceGroupName = Create-ResourceGroup $location
+	# $targetResourceGroupName = Create-ResourceGroup $location 1
+
+	$location = "centraluseuap"
+	$resourceGroupName = "iaasvm-pstest-rg"
+	$targetResourceGroupName = "hiagarg"
+
+	$vaultName = "iaasvm-pstest-vault"
+	$vmName = "iaasvm-pstest-vm"
+
+	$saName = "pstestsa2"
 
 	try
 	{
 		# Setup
-		$saName = Create-SA $resourceGroupName $location
-		$vm = Create-VM $resourceGroupName $location
-		$vault = Create-RecoveryServicesVault $resourceGroupName $location
+		# $saName = Create-SA $resourceGroupName $location
+
+		$vm = Get-AzVM -ResourceGroupName $resourceGroupName -Name $vmName # Create-VM $resourceGroupName $location
+		$vault = Get-AzRecoveryServicesVault -ResourceGroupName $resourceGroupName -Name $vaultName # Create-RecoveryServicesVault $resourceGroupName $location
 		Set-AzRecoveryServicesVaultProperty -VaultId $vault.ID -SoftDeleteFeatureState "Disable"
-		$item = Enable-Protection $vault $vm
+
+		$item = Get-AzRecoveryServicesBackupItem -BackupManagementType AzureVM -WorkloadType AzureVM -VaultId $vault.ID # Enable-Protection $vault $vm
 		$backupJob = Backup-Item $vault $item
 		$rp = Get-RecoveryPoint $vault $item $backupJob
 
@@ -602,8 +684,8 @@ function Test-AzureVMFullRestore
 	finally
 	{
 		# Cleanup
-		Cleanup-ResourceGroup $resourceGroupName
-		Cleanup-ResourceGroup $targetResourceGroupName
+		# Cleanup-ResourceGroup $resourceGroupName
+		# Cleanup-ResourceGroup $targetResourceGroupName
 	}
 }
 
@@ -699,7 +781,7 @@ function Test-AzureVMBackup
 	try
 	{
 		# Setup
-		$vault = $vault = Get-AzRecoveryServicesVault -ResourceGroupName $resourceGroupName -Name $vaultName
+		$vault = Get-AzRecoveryServicesVault -ResourceGroupName $resourceGroupName -Name $vaultName
 				
 		Set-AzRecoveryServicesVaultProperty -VaultId $vault.ID -SoftDeleteFeatureState "Disable"
 		
@@ -720,14 +802,19 @@ function Test-AzureVMBackup
 
 function Test-AzureVMSetVaultContext
 {
-	$location = "southeastasia"
-	$resourceGroupName = Create-ResourceGroup $location
+	# $location = "eastasia"
+	# $resourceGroupName = Create-ResourceGroup $location
+
+	$location = "centraluseuap"
+	$resourceGroupName = "iaasvm-pstest-rg"
+	$vaultName = "iaasvm-pstest-vault"
+	$vmName = "iaasvm-pstest-vm"
 
 	try
 	{
 		# Setup
-		$vm = Create-VM $resourceGroupName $location
-		$vault = Create-RecoveryServicesVault $resourceGroupName $location
+		$vm = Get-AzVM -ResourceGroupName $resourceGroupName -Name $vmName # Create-VM $resourceGroupName $location
+		$vault = Get-AzRecoveryServicesVault -ResourceGroupName $resourceGroupName -Name $vaultName # Create-RecoveryServicesVault $resourceGroupName $location
 
 		# disable soft delete for successful cleanup
 		Set-AzRecoveryServicesVaultProperty -VaultId $vault.ID -SoftDeleteFeatureState "Disable"
@@ -736,6 +823,10 @@ function Test-AzureVMSetVaultContext
         Start-TestSleep 5000
 
 		Set-AzRecoveryServicesVaultContext -Vault $vault | Out-Null
+
+		$item = Get-AzRecoveryServicesBackupItem -BackupManagementType AzureVM -WorkloadType AzureVM -VaultId $vault.ID
+
+		Disable-AzRecoveryServicesBackupProtection -Item $item -RemoveRecoveryPoints -Force
 
 		# Get default policy
 		$policy = Get-AzRecoveryServicesBackupProtectionPolicy `
@@ -747,24 +838,26 @@ function Test-AzureVMSetVaultContext
 			-Name $vm.Name `
 			-ResourceGroupName $vm.ResourceGroupName;
 
-		$container = Get-AzRecoveryServicesBackupContainer `
-			-ContainerType AzureVM `
-			-Status Registered;
+		# $container = Get-AzRecoveryServicesBackupContainer `
+		# 	-ContainerType AzureVM `
+		# 	-Status Registered;
+		# 
+		# $item = Get-AzRecoveryServicesBackupItem `
+		# 	-Container $container `
+		# 	-WorkloadType AzureVM
 
-		$item = Get-AzRecoveryServicesBackupItem `
-			-Container $container `
-			-WorkloadType AzureVM
+		# $item = Get-AzRecoveryServicesBackupItem -BackupManagementType AzureVM -WorkloadType AzureVM -VaultId $vault.ID
 
 		# Disable protection
-		Disable-AzRecoveryServicesBackupProtection `
-			-Item $item `
-			-RemoveRecoveryPoints `
-			-Force;
+		# Disable-AzRecoveryServicesBackupProtection `
+		# 	-Item $item `
+		# 	-RemoveRecoveryPoints `
+		# 	-Force;
 	}
 	finally
 	{
 		# Cleanup
-		Cleanup-ResourceGroup $resourceGroupName
+		# Cleanup-ResourceGroup $resourceGroupName
 	}
 }